# server/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GroChain Backend API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for GroChain backend."
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Listing": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "product": { "type": "string" },
          "price": { "type": "number" },
          "quantity": { "type": "number" },
          "farmer": { "type": "string" },
          "partner": { "type": "string" },
          "images": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["active","sold","removed"] }
        }
      },
...
    }
  },
  "security": [ { "bearerAuth": [] } ],
  "paths": {
...
    "/api/marketplace/listings": { "get": { "summary": "List marketplace listings", "responses": { "200": { "description": "OK" } } }, "post": { "summary": "Create listing", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Listing" } } } }, "responses": { "201": { "description": "Created" } } } },
    "/api/marketplace/listings/{id}": { "get": { "summary": "Get listing", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } } }, "patch": { "summary": "Update listing", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/listings/{id}/unpublish": { "patch": { "summary": "Unpublish listing", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/orders": { "post": { "summary": "Create order", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "buyer": {"type":"string"}, "items": {"type":"array","items":{"type":"object","properties":{"listing":{"type":"string"},"quantity":{"type":"integer"}}}} }, "required": ["buyer","items"] } } } }, "responses": { "201": { "description": "Created" } } } },
    "/api/marketplace/orders/{id}": { "get": { "summary": "Get order details", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } }, "patch": { "summary": "Update order status", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "status": {"type":"string"} } } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/orders/{id}/cancel": { "patch": { "summary": "Cancel order", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/orders/{id}/tracking": { "get": { "summary": "Track order", "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/search-suggestions": { "get": { "summary": "Search suggestions", "parameters": [ {"name":"q","in":"query","schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/favorites/{userId}": { "get": { "summary": "Get favorites", "parameters": [ {"name":"userId","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } } },
    "/api/marketplace/favorites": { "post": { "summary": "Add to favorites", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "userId": {"type":"string"}, "listingId": {"type":"string"} }, "required": ["userId","listingId"] } } } }, "responses": { "201": { "description": "Created" } } } },
    "/api/marketplace/favorites/{userId}/{listingId}": { "delete": { "summary": "Remove from favorites", "parameters": [ {"name":"userId","in":"path","required":true,"schema":{"type":"string"}}, {"name":"listingId","in":"path","required":true,"schema":{"type":"string"}} ], "responses": { "200": { "description": "OK" } } } },
...
  }
}